Add a shim around OutputManager to avoid Rich imports
